remove askama.toml and just use base templates
this is because the build script put the minified html into the
templates directory itself, and that would cause the build script to be
reran every single time.

The best fix for this would be to have askama read from OUT_DIR, instead
of CARGO_MANIFEST_DIR, but the creator thinks that
this is not a good use case https://github.com/djc/askama/pull/472. Therefore,
I'll probably just move to another templating engine.